requirements.txt
Flask
Flask-Cors
Flask-restful
krbcontext
pywinrm
pywinrm[kerberos]
pywinrm[credssp]
